pyproject.toml
[tool.poetry]
name = "bot-client"
version = "0.1.0"
description = "This bot client will ran in a docker environment, and whenever requested, it will generate an Agent that will join a given chat to interact with a user"
authors = ["ismael-su <ismael@rumble.studio>"]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.7.0"
torch = "^1.9.0"
asyncio = "^3.4.3"
python-socketio = "^5.3.0"
fastapi = "^0.65.2"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
pydantic = "^1.8.2"
numpy = "^1.20.3"
datasets = "^1.8.0"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
